Move dialog to file
