+++ material/build.gradle
@@ -1,7 +1,10 @@
 import dependencies.Dependencies
 import dependencies.Versions
 
-apply plugin: 'com.android.library'
+plugins {
+    id 'com.android.library'
+    id 'kotlin-android'
+}
 
 apply from: '../config/quality.gradle'
